Move demo to app directory
