+++ demo/src/app/form.ts
@@ -1,0 +1,17 @@
+import { RadioGroup, Radio } from 'components/radio'
+import Switch from 'components/switch'
+import { Fieldset, Legend } from 'components/fieldset'
+import LiveCodeSample from 'components/live-code-sample'
+import { SheetThemeMode } from 'components/theme-mode'
+import { DetentsSelector } from 'components/detents-selector'
+
+export const form = {
+  RadioGroup,
+  Radio,
+  Switch,
+  Fieldset,
+  Legend,
+  LiveCodeSample,
+  SheetThemeMode,
+  DetentsSelector
+}
