--- demo/src/components/meta.tsx
@@ -1,70 +1,0 @@
-import React from 'react'
-import Head from 'next/head'
-
-const Meta: React.FC<{}> = () => {
-  return (
-    <Head>
-      <title>react-sheet-slide</title>
-      <meta charSet="utf-8" />
-      <meta
-        name="viewport"
-        content="width=device-width, initial-scale=1, viewport-fit=cover"
-      />
-      <link rel="icon" href="/react-sheet-slide/favicon.ico" />
-      <meta name="apple-mobile-web-app-capable" content="yes" />
-      <meta name="apple-mobile-web-app-status-bar-style" content="default" />
-      <link
-        rel="manifest"
-        href="/react-sheet-slide/manifest.webmanifest"
-        crossOrigin="anonymous"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="32x32"
-        href="/react-sheet-slide/icons/apple-32x32.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="48x48"
-        href="/react-sheet-slide/icons/apple-48x48.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="72x72"
-        href="/react-sheet-slide/icons/apple-72x72.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="96x96"
-        href="/react-sheet-slide/icons/apple-96x96.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="144x144"
-        href="/react-sheet-slide/icons/apple-144x144.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="192x192"
-        href="/react-sheet-slide/icons/apple-192x192.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="256x256"
-        href="/react-sheet-slide/icons/apple-256x256.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="384x384"
-        href="/react-sheet-slide/icons/apple-384x384.png"
-      />
-      <link
-        rel="apple-touch-icon"
-        sizes="512x512"
-        href="/react-sheet-slide/icons/apple-512x512.png"
-      />
-    </Head>
-  )
-}
-
-export default Meta
