+++ demo/src/types/index.d.ts
@@ -1,7 +1,7 @@
 declare global {
   var __isDev__: boolean
 
-  type Nothing = Record<never. never>
+  type Nothing = Record<never, never>
 }
 
 export {}
