+++ demo/tailwind.config.js
@@ -1,7 +1,7 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   corePlugins: {
-    preflight: false
+    preflight: true
   },
   content: ["./src/**/*.{js,ts,jsx,tsx,mdx}"],
   theme: {
